pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "migres"
version = "0.1.1.dev1"
description = "Database migration tool from for transfering database from MariaDB to PostgreSQL/Supabase"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Mayowa Ogungbola", email = "ogungbolamayowa@email.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Database",
    "Topic :: Utilities",
]
keywords = ["migration", "database", "mariadb", "postgresql", "supabase"]
dependencies = [
    "python-dotenv>=0.19.0",
    "pydantic>=1.8.0,<2.0.0",
    "pymysql>=1.0.0",
    "psycopg2-binary>=2.9.0",
    "pandas>=1.3.0",
    "configparser>=5.0.0",
]

[project.urls]
Homepage = "https://github.com/Phenzic/migres"
Documentation = "https://github.com/Phenzic/migres#readme"
Issues = "https://github.com/Phenzic/migres/issues"
Source = "https://github.com/Phenzic/migres"

[project.scripts]
migres = "migres.cli:main"

[tool.setuptools.packages]
find = {}